Introduces ValueVariableNumber

+++ platform/runner/plugin/std/sys/src/main/kotlin/io/hamal/plugin/std/sys/func/FuncDeployFunction.kt
@@ -33,7 +33,7 @@ class FuncDeployFunction(
                 null
             }
             val version = if (arg1.type("version") == ValueNumber::class) {
-                CodeVersion(arg1.getNumber("version").intValue)
+                CodeVersion(arg1.getNumber("version"))
             } else {
                 null
             }
